Module edgequest::core::creature

A creature holds a state, an AI, some Stats, and an Actor

The intention is that Creature binds all neccesary parts that an AI needs to function such that each part is easily passed in and also easily replacable

Re-exports

pub use self::actions::Actions;

pub use self::actor::Actor;

pub use self::stats::Stats;

Modules

actions
Enum representing possible actions creatures can take

actor
An Actor is an object that is renderable to the screen and is responsible for moving itself

ai
Metapackage to extend an interface to ai

stats
Stats is a struct intended to hold a set of mutable data that describes the statistics of Creatures

Structs

Creature
Creature holds a Actor and an AI, basically a package that we can create monsters from