Trait edgequest::core::creature::ai::AI

pub trait AI {
    fn take_turn(
        &mut self, 
        map: &Grid<Tile>, 
        player: &Creature, 
        me: &mut Actor, 
        stats: &mut Stats
    ) -> Actions;

    fn box_clone(&self) -> Box<dyn AI>;

    fn is_oob(&mut self, x: isize, y: isize, map: &Grid<Tile>) -> bool { ... }
}

Represents basic actions AI can take in the game

An AI is a trait because we want all AI to follow a similar pattern and thus be Boxable and able to be given to Creatures. Thus, all AI patterns are trait objects.

Required Methods

fn take_turn(
&mut self,
map: &Grid<Tile>,
player: &Creature,
me: &mut Actor,
stats: &mut Stats
) -> Actions

Make the AI take it's turn based on map, player, and itself

NOTE: AIs are basically just state deciders at this point but more complex AIs have to be state machines in of themselves in order to create complex behaviors. At some point they should take in a state, a vector of all creatures on the floor (for monster infighting, fight-flight) and maybe even some sort of "mood" though that would be a part of the Creature. I am completely considering adding randomized personalities to monsters to create even more combinations of behavior.

fn box_clone(&self) -> Box<dyn AI>

Allow boxed trait objects to be cloned

Provided Methods

fn is_oob(&mut self, x: isize, y: isize, map: &Grid<Tile>) -> bool

Determine if the AI has gone out of bounds with respect to the given map