Struct edgequest::core::init::config::Config

pub struct Config {
    pub screen_width: isize,
    pub screen_height: isize,
    pub map_width: isize,
    pub map_height: isize,
    pub fullscreen: bool,
    pub fontpath: String,
    pub fonttype: FontType,
    pub fontlayout: FontLayout,
    pub renderer: Renderer,
}

A struct to hold data gathered from a config.yml file. You should not need to create your own, instead, get a filled out struct from load()

What data should be in your configuration file

See load()

Determining font settings

Firstly, pick a font from the fonts directory. Once you have a font you like, add it to the fontpath in the configuration file, so it looks like fontpath: fonts/yourfont.png.

The next step is figuring out your FontType and FontLayout. Your font file will be in the form name_FontType_FontLayout. Here is how they line up.

FontType

aa - Default

gs - Greyscale

FontLayout

tc - Tcod

ro - AsciiInRow

as - AsciiInCol

Fields
screen_width: isize screen_height: isize map_width: isize map_height: isize fullscreen: bool fontpath: String fonttype: FontType fontlayout: FontLayout renderer: Renderer