Struct edgequest::core::object::fighter::Fighter

pub struct Fighter {
    pub pos: Pos,
    // some fields omitted
}

A Fighter is something that can fight, but also be rendered to the screen

NOTE: From that description, I think something might have to change...

Fields
pos: Pos
Methods

impl Fighter

pub fn move_cart(&mut self, x: isize, y: isize)

Move the Entity by x in the x direction and y in the y direction.

This does not overwrite the positon, only add to it. x and y can be positive or negative.

x - ammount to move in the x direction

y - ammount to move in the x direction

pub fn move_pos(&mut self, pos: Pos)

Move the Entity by adding a new Pos to it

This does not overwrite the positon, only add to it. If values in Pos are negative, this will then just subtract the appropriate values.

pos - Pos struct of ammount to move in both x and y directions

pub fn new(
name: &'static str,
glyph: char,
pos: (isize, isize),
fg: (u8, u8, u8),
bg: (u8, u8, u8)
) -> Fighter

Return a new Entity

pub fn set_pos(&mut self, pos: Pos)