Struct edgequest::core::object::pos::Pos

pub struct Pos {
    pub x: isize,
    pub y: isize,
}

Hold an x, y cartesian coordinate. Should be used when explicitly needing to manipulate points, not just store them. Length and width should be stored as tuples, however a monster should be a pos to calculate distance between it and the player.

x - x axis location y - y axis location

Fields
x: isize y: isize
Methods

impl Pos

pub fn new(x: isize, y: isize) -> Pos

Return a new Pos

pub fn from_usize(x: usize, y: usize) -> Pos

Return a new Pos from usize's

pub fn from_tup(pos: (isize, isize)) -> Pos

Return a Pos from a tuple

pub fn as_tup(&self) -> (isize, isize)

Return a tuple from a Pos

pub fn origin() -> Pos

Return a Pos at the origin (0, 0)