Struct edgequest::core::renderer::Renderer

pub struct Renderer {
    pub sc_debug: bool,
    // some fields omitted
}

Helper for rendering things to the screen

Tracks the player and automatically scrolls the screen around to match where they go. This will never try to draw things outside of the given dimensions due to the way it handles determining whether something should be drawn or not.

Fields
sc_debug: bool
Methods

impl Renderer

pub fn debug_render_scent_map(&mut self, con: &mut Console, dungeon: &Dungeon)

Render scent as a fine, purple mist, just like real life

pub fn draw_all(&mut self, con: &mut Console, world: &World)

Draw all.

You'll have to render this console to root (unless you passed root in) and always flush() the root console.

pub fn draw_creature(
&self,
con: &mut Console,
pos: Pos,
ce: &Entity,
world: &World
)

Draw creatures with "transparent" backgrounds

NOTE: Way too similar to draw_entity(). Plus, the whole bg thing seems weird. Better way to do it?

pub fn draw_entity(&self, con: &mut Console, pos: Pos, ce: &Entity)

Put an Entity on the console

con - Tcod Console

entity - Entity

pub fn new(map: (isize, isize), screen: (isize, isize)) -> Renderer

Return a new Renderer

map - Pos that holds the map dimensions

screen - Pos that holds the screen dimensions