Type Definition edgequest::core::world::dungeon::map::grid::Grid

type Grid<T> = Vec<Vec<T>>;

Hold a 2D vector of Ts

Primarily used for dungeon generation and cellular automatons, also it looks better than Vec<Vec<T>> and is more intuitive as to what it actually represents