Struct edgequest::core::world::dungeon::Dungeon

pub struct Dungeon {
    pub width: usize,
    pub height: usize,
    pub grid: Grid<Tile>,
}

Dungeon struct to stitch together all builders and cellular automatons

Fields
width: usize height: usize grid: Grid<Tile>
Methods

impl Dungeon

pub fn build(&mut self)

Make the dungeon

pub fn get_valid_location(grid: &Grid<Tile>) -> (usize, usize)

Get the player's starting location as a Pos

NOTE: Should be deprecated and removed once stairs show up

pub fn is_valid(&self, x: usize, y: usize) -> bool

Check to see if a specific tile is valid, i.e. walkable and in the map bounds

pub fn new(map_dim: (usize, usize)) -> Dungeon

Return a new Dungeon