Struct edgequest::core::world::World

pub struct World {
    pub player: Creature,
    pub floor: Floor,
    pub floor_stack: Vec<Floor>,
    pub floor_num: usize,
    pub tcod_map: Map,
}

Fields
player: Creature floor: Floor floor_stack: Vec<Floor> floor_num: usize tcod_map: Map
Methods

impl World

pub fn test_empty(&mut self)

Empty out the floor

pub fn is_valid_pos(&self, x: isize, y: isize) -> bool

Check to see if a specific position is valid, i.e. tile::walkable and in the map bounds

pub fn check_trap(&mut self)

Check to see if a tile is a trap

Should only be called after checking tile validity to avoid OOB errors

pub fn get_bg_color_at(&self, pos: Pos) -> RGB

Return the bg color of a tile at a point

NOTE: Clearly does not give a fuck if you go OOB, probably should change Interestingly, it's only used by the renderer and that can't display stuff out of bounds (thanks, camera) so maybe it's not important to have needless code?

pub fn go_down(&mut self)

Go downstairs if possible

pub fn go_up(&mut self)

Save the current floor and go up one floor

pub fn player_go_up(&mut self)

See if the player is able to go up on the current tile and draw some stuff to the log

pub fn player_go_down(&mut self)

See if the player is able to go down on the current tile and draw some stuff to the log

pub fn test_traverse(&mut self)

Temporary function for stair traversal. In the future floors will need to be saved

pub fn new_tcod_map(map_dim: Pos, dungeon: &Dungeon) -> Map

Return a tcod map based on dungeon features (Essentially what walls you can walk and see through)

pub fn new(map_dim: Pos) -> Self

Return a new World

pub fn get_mut_tile_at(&mut self, x: isize, y: isize) -> &mut Tile

pub fn get_tile_at(&self, x: isize, y: isize) -> &Tile

pub fn update_fov(&mut self)

Update the fov map from the player's perspective

pub fn update(&mut self)

Update the game world