Struct edgequest::dungeon::Dungeon

pub struct Dungeon {
    pub grid: Vec<Vec<u8>>,
    pub w: i32,
    pub h: i32,
    // some fields omitted
}

Dungeon struct to generate a bitmap dungeon (1s and 0s)

Q: Why not use booleans? A: Eventually this is going to have more than just 1s and 0s.

grid - 2D Vec of u8s

w - Width of desired map

h - Height of desired map

rooms - Vec of Rects but since Rect is private you should be getting Dungeons via new()

Fields
grid: Vec<Vec<u8>> w: i32 h: i32
Methods

impl Dungeon

fn new(w: i32, h: i32, rooms: i32) -> Dungeon

Get a new dungeon

fn regen(&mut self)

Regenerate the dungeon map

fn get_starting_location(&self) -> Pos

Get the player's starting location as a Pos