Struct edgequest::game::Game

pub struct Game {
    pub player: Entity,
    pub map: Map,
    pub dungeon: Dungeon,
}

Game struct. Holds a player and a map

player - Entity to represent the player

map - Map object to represent the current floor

Fields
player: Entity map: Map dungeon: Dungeon
Methods

impl Game

fn new_player() -> Entity

Return a new player Entity

fn new_map(width: i32, height: i32) -> Map

Return a new empty Map

This function assumes you will just be passing in tcod::console::Root.width() and height(), so inputs are i32s instead of usizes (they get converted)

fn new_dungeon(width: i32, height: i32) -> Dungeon

fn new(width: i32, height: i32) -> Game

Return a new Game

This function assumes you will just be passing in tcod::console::Root.width() and height(), so inputs are i32s instead of usizes (they get converted)

fn draw_all(&mut self, con: &mut Console)

Draw all.

You'll have to render this console to root (unless you passed root in) and always flush() the root console.

fn capture_keypress(&mut self, root: &mut Root)

Capture keyboard input and process it with tcod