Struct edgequest::core::world::World

pub struct World {
    pub player: Fighter,
    pub cur_dungeon: Dungeon,
    pub creatures: Vec<Box<Creature>>,
    pub dungeon_stack: Vec<Dungeon>,
}

Fields
player: Fighter cur_dungeon: Dungeon creatures: Vec<Box<Creature>> dungeon_stack: Vec<Dungeon>
Methods

impl World

pub fn test_empty(&mut self)

Empty out the floor

pub fn get_bg_color_at(&self, x: usize, y: usize) -> RGB

Return the bg color of a tile at a point

NOTE: Clearly does not give a fuck if you go oob, probably should change

pub fn can_go_down(&self) -> bool

Go downstairs

pub fn go_down(&mut self)

pub fn can_go_up(&self) -> bool

Go upstairs

pub fn go_up(&mut self)

pub fn test_traverse(&mut self)

Temporary function for stair traversal. In the future floors will need to be saved.

pub fn new(map_dim: (isize, isize)) -> World

Return a new World

pub fn update(&mut self)

Update the game world